requirements.txt
# Remove CUDA-specific index URL since we're on macOS
# --extra-index-url https://download.pytorch.org/whl/cu118

# Core ML and data processing
torch>=2.2.0  # Latest stable version for macOS
torchvision>=0.17.0
torchaudio>=2.2.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0

# Deep learning and transformers
transformers>=4.36.0
sentence-transformers>=2.2.0
accelerate>=0.25.0
datasets>=2.14.0
evaluate>=0.4.0
tokenizers>=0.14.0
safetensors>=0.4.0

# Visualization and analysis
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0
bokeh>=3.3.0

# Jupyter and development
jupyter>=1.0.0
ipykernel>=6.0.0
ipywidgets>=8.0.0
notebook>=7.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.66.0
openpyxl>=3.1.0
tabulate>=0.9.0

# Text processing
langdetect>=1.0.9
regex>=2023.0.0

# Azure services
azure-ai-textanalytics>=5.3.0
azure-core>=1.28.0

# Additional ML tools
setfit>=1.0.0
embetter[sentence-tfm]>=0.6.0
hiclass>=5.0.0
skops>=0.11.0

# OpenAI
openai>=1.0.0

accelerate==1.3.0
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
anywidget==0.9.13
asttokens==3.0.0
attrs==25.1.0
azure-common==1.1.28
bokeh>=3.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
decorator==5.1.1
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
et-xmlfile==2.0.0
executing==2.2.0
filelock==3.17.0
fonttools==4.55.7
frozenlist==1.5.0
fsspec==2024.9.0
fsspec[http]==2024.9.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.0
idna==3.10
ipython==8.31.0
isodate==0.7.2
jedi==0.19.2
jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jupyter-scatter==0.20.0
jupyterlab-widgets==3.0.13
kiwisolver==1.4.8
markupsafe==3.0.2
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.61.0
openpyxl==3.1.5
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
pooch==1.8.2
prompt-toolkit==3.0.50
propcache==0.2.1
psutil==6.1.1
psygnal==0.11.1
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==19.0.0
pydantic-core==2.27.2
pydantic==2.10.6
pygments==2.19.1
pynndescent==0.5.13
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyyaml==6.0.2
radicli==0.0.25
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
sentence-transformers==3.4.0
sentence-transformers[train]==3.4.0
setfit==1.1.1
setuptools==75.8.0
six==1.17.0
skops==0.11.0
sniffio==1.3.1
stack-data==0.6.3
sympy==1.13.1
tabulate==0.9.0
threadpoolctl==3.5.0
tokenizers==0.21.0
torchaudio>=2.2.0
torchvision>=0.17.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
traittypes==0.2.1
transformers==4.48.1
typing-extensions==4.12.2
tzdata==2025.1
umap-learn==0.5.7
umap==0.1.1
urllib3==2.3.0
wasabi==1.1.3
wcwidth==0.2.13
widgetsnbextension==4.0.13
xxhash==3.5.0
yarl==1.18.3
